+++ canvas_editor/account_management/forms/register_form.py
@@ -1,11 +1,9 @@
-from canvas import message_dict
-
+from typing import Any
 
 from django import forms
 from django.contrib.auth.models import User
 
-
-from typing import Any
+from canvas import message_dict
 
 
 class RegisterForm(forms.Form):
